+++ Docs/CHANGELOG.md
@@ -7,13 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.6.180] - 2025-05-12
 ### Changed
 - Updated the menu in TgDownloaderConsole
-- Updated license view in TgDownloaderConsole
-- Updated license view in OpenTgResearcherDesktop
+- Updated license view
 - Updated NuGet packages
 ### Added
 - Added methods to memorize window size and position between sessions in OpenTgResearcherDesktop
-- Added license request
 - Added license clear
+- Added license request
 - Increased the limit on the number of download threads for a user with a paid or test license
 
 ## [0.6.140] - 2025-04-26